+++ cli/cd.c
@@ -4,17 +4,12 @@
 #include "aux/toggle.h"
 #include "aux/config.h"
 
-/*
- * TODO:
- * 1. Use conts agrv. Gotta create a new dynamic struct cuz code alters it
- */
-
-int tec_cli_cd(int argc, char **argv, tec_ctx_t *ctx)
+int tec_cli_cd(int argc, const char **argv, tec_ctx_t *ctx)
 {
     tec_arg_t args;
+    tec_argvec_t argvec;
     const char *errfmt;
     int c, i, retcode, status;
-    char alias[IDSIZ + 1] = { 0 };
     int opt_quiet, opt_help, opt_cd_dir, opt_cd_toggle;
 
     retcode = LIBTEC_OK;
@@ -22,7 +17,10 @@ int tec_cli_cd(int argc, char **argv, tec_ctx_t *ctx)
     opt_cd_toggle = opt_cd_dir = true;
     errfmt = "cannot switch to '%s': %s";
     args.env = args.desk = args.taskid = NULL;
-    while ((c = getopt(argc, argv, ":d:e:hnqN")) != -1) {
+
+    argvec_init(&argvec);
+    argvec_parse(&argvec, argc, argv);
+    while ((c = getopt(argvec.count, argvec.argv, ":d:e:hnqN")) != -1) {
         switch (c) {
         case 'd':
             args.desk = optarg;
@@ -61,20 +59,20 @@ int tec_cli_cd(int argc, char **argv, tec_ctx_t *ctx)
         return status;
 
     /* Check that alias '-' is not passed with other task IDs nor duplicated.  */
-    for (int idx = i; idx < argc; ++idx) {
-        if (strcmp(argv[idx], "-") == 0 && argc - i > 1)
+    for (int idx = i; idx < argvec.count; ++idx) {
+        if (strcmp(argvec.argv[idx], "-") == 0 && argvec.count - i > 1)
             return elog(1, "alias '-' is used alone");
     }
 
     /* Resolve alias '-' to switch to previous task ID.  */
-    if (argv[i] && strcmp("-", argv[i]) == 0) {
+    if (argvec.argv[i] && strcmp("-", argvec.argv[i]) == 0) {
         if ((status = toggle_task_get_prev(teccfg.base.task, &args)))
             return elog(1, errfmt, "PREV", "no previous task ID");
-        argv[i] = strncpy(alias, args.taskid, IDSIZ);
+        argvec_replace(&argvec, i, args.taskid, IDSIZ);
     }
 
     do {
-        args.taskid = argv[i];
+        args.taskid = argvec.argv[i];
         if ((status = check_arg_task(&args, errfmt, opt_quiet))) {
             ;
         } else if ((status = hook_action(&args, "cd"))) {
@@ -87,7 +85,11 @@ int tec_cli_cd(int argc, char **argv, tec_ctx_t *ctx)
             }
         }
         retcode = status == LIBTEC_OK ? retcode : status;
-    } while (++i < argc);
+    } while (++i < argvec.count);
 
-    return retcode == LIBTEC_OK && opt_cd_dir ? tec_pwd_task(&args) : retcode;
+    if (retcode == LIBTEC_OK && opt_cd_dir)
+        retcode = tec_pwd_task(&args) == LIBTEC_OK ? retcode : status;
+
+    argvec_free(&argvec);
+    return retcode;
 }