+++ cli/init.c
@@ -1,7 +1,7 @@
 #include "tec.h"
 #include "aux/config.h"
 
-int tec_cli_init(int argc, char **argv, tec_ctx_t *ctx)
+int tec_cli_init(int argc, const char **argv, tec_ctx_t *ctx)
 {
     int status;